+++ R/DBI.R
@@ -17,6 +17,8 @@
 #' @examples
 #' # You use a dbPool in the same way as a standard DBI connection
 #' pool <- dbPool(RSQLite::SQLite())
+#' pool
+#'
 #' DBI::dbWriteTable(pool, "mtcars", mtcars)
 #' dbGetQuery(pool, "SELECT * FROM mtcars LIMIT 4")
 #'